+++ src/Urls/UrlHelper.php
@@ -6,6 +6,7 @@ namespace Thinktomorrow\Chief\Urls;
 
 use Illuminate\Support\Collection;
 use Illuminate\Database\Eloquent\Model;
+use Thinktomorrow\Chief\Management\Managers;
 use Thinktomorrow\Chief\Common\Helpers\Memoize;
 use Thinktomorrow\Chief\Concerns\Morphable\Morphables;
 use Thinktomorrow\Chief\FlatReferences\FlatReferencePresenter;
@@ -31,13 +32,12 @@ class UrlHelper
      * Internal api for fetching all models.
      * This will return a grouped values array ready for select fields
      *
-     * @param bool $onlySingles
      * @param Model|null $ignoredModel
      * @return array
      */
-    public static function allModelsWithoutSelf(Model $ignoredModel = null): array
+    public static function allModelsExcept(Model $ignoredModel = null, bool $online = false): array
     {
-        $models = static::models(false, $ignoredModel, false);
+        $models = static::models(false, $ignoredModel, $online);
 
         return FlatReferencePresenter::toGroupedSelectValues($models)->toArray();
     }
@@ -61,15 +61,41 @@ class UrlHelper
     }
 
 
-    public static function models(bool $onlySingles = false, Model $ignoredModel = null, $online = true)
+    public static function models(bool $onlySingles = false, Model $ignoredModel = null, bool $online = true)
     {
-        $models = chiefMemoize('all-online-models', function () use ($onlySingles, $online) {
+        $types = [];
+
+        if ($onlySingles) {
+            $types = ['singles'];
+        }
+
+        return self::modelsByType($types, $ignoredModel, $online);
+    }
+
+    public static function modelsByKeys(array $keys, Model $ignoredModel = null, bool $online = true)
+    {
+        $managers = app(Managers::class);
+
+        $whitelistedDatabaseTypes = [];
+
+        foreach ($keys as $key) {
+            $manager = $managers->findByKey($key);
+            $whitelistedDatabaseTypes[] = $manager->modelInstance()->getMorphClass();
+        }
+
+        return static::modelsByType($whitelistedDatabaseTypes, $ignoredModel, $online);
+    }
+
+
+    public static function modelsByType(array $types, Model $ignoredModel = null, bool $online = true)
+    {
+        $models = chiefMemoize('all-online-models', function () use ($types, $online) {
             $builder = UrlRecord::whereNull('redirect_id')
                 ->select('model_type', 'model_id')
                 ->groupBy('model_type', 'model_id');
 
-            if ($onlySingles) {
-                $builder->where('model_type', 'singles');
+            if (!empty($types)) {
+                $builder->whereIn('model_type', $types);
             }
 
             return $builder->get()->mapToGroups(function ($record) {
@@ -83,7 +109,7 @@ class UrlHelper
 
                 return is_null($model);
             })->flatten();
-        }, [$onlySingles]);
+        });
 
         if ($ignoredModel) {
             $models = $models->reject(function ($model) use ($ignoredModel) {