+++ tests/Feature/Management/Fakes/ManagedModelFakeFirst.php
@@ -24,8 +24,6 @@ use \Astrotomic\Translatable\Translatable as BaseTranslatable;
 
 class ManagedModelFakeFirst extends Model implements ManagedModel, TranslatableContract, HasAsset, ActsAsParent, ActsAsChild, StatefulContract
 {
-    private $current_state = 'draft';
-
     use HasDynamicAttributes {
         HasDynamicAttributes::fill as hasDynamicAttributesFill;
         HasDynamicAttributes::getAttribute as private hasDynamicAttributesGetAttribute;
@@ -113,12 +111,12 @@ class ManagedModelFakeFirst extends Model implements ManagedModel, TranslatableC
 
     public function stateOf($key): string
     {
-        return $this->current_state;
+        return $this->$key;
     }
 
     public function changeStateOf($key, $state)
     {
-        $this->current_state = $state;
+        $this->$key = $state;
     }
 
     public function fill(array $attributes)